requirements.txt
fastapi
uvicorn[standard]
python-multipart
aiofiles
langchain==0.1.16
langchain-community>=0.0.10
langchain-openai>=0.0.8
openai>=1.10.0,<2.0.0
chromadb==0.4.22
pypdf==4.0.1
python-dotenv==1.0.0
tiktoken==0.5.2
sentence-transformers==2.2.2
pandas
numpy
tabula-py
pdfplumber 